feat(core): bundle the search schema as a mountable declaration module
The prebuilt LDE images - @lde/search-indexer and @lde/search-api-server - boot
by loading a module that default-exports a deployment's search type declarations
as plain data. Adopting either means having one, so build it first: it is the
prerequisite both share, and the step that unblocks generating the extraction
CONSTRUCTs from the schema.

A mounted file cannot resolve bare specifiers, so module.ts is bundled to
importless ESM with @lde/search aliased to a stub. defineSearchType is the
identity function at runtime, so nothing is lost and the real package's CJS
dependencies stay out of what must be a lean module; the derive functions
survive as the plain functions they are.

Splits declarations from construction: declarations.ts holds the SearchType
literals, schema.ts builds SEARCH_SCHEMA from them and re-exports. The bundle
would otherwise drag in the searchSchema call, which a plain-data module must
not reach.

The contract is now printed through the mounted module with search-print-sdl,
the writer @lde/search-api-graphql ships, rather than from code - so the module
is load-bearing and cannot rot unnoticed: if it stops loading, the contract stops
printing. The generated SDL is byte-identical, which is the evidence that the
module describes the same API.
